Breath Life into Cortex and Measure Time to Load
Declare Emacs bankruptcy and revive Cortex!
Set header arguments for the complete file. Tangle all source blocks to
the init.el file in a position where Emacs will look for it.
This config shall work cross platform on Linux, macOS and Windows
systems in the end. MAke sure that ~ is pointing to the users correct
home folder.
Save the time of the start of Emacs in order to monitor how long Cortex
needs to load. Print that info into the *Messages* buffer.