Start Emacs in server mode and Windows setup docu
Attaching files to an already started Emacs session is the sane way of
dealing with multiple files from various sources. This saves a lot of
time especially on Windows systems. The default behavior there is to
open a file selected in the file browser in its own, isolated instance
of Emacs. The server mode enables to attach the file to the server
session without the need to reload the complete Emacs configuration.

Setting up Emacs on a Windows host is documented.