Minor change to minibuffer completion, noweb-style
Make completion in minibuffer ignore case and use Icomplete for
visualisation of candidates.

This is a first try of using a noweb approach to the config.
I think it adds to the readability of Cortex. I can now potentially
comment on every variable change individually right next to the code. I
had to do this in a big block of documentation sometimes far away from
the actual line that implements the change.