Use orderless completion, limit org-agenda-files
Orderless completion allows for quick narrowing of candidates even if I
don't exactly know what I'm searching for. This makes e.g. using tags in
filenames work when finding files.

In order to have a fast loading org-agenda, I limit my files to scan to
the main todo files. This is not much of a problem now, but can be seen
as a proactive measure.
This impacts how I structure my todos: They should not be part of the
result of the task. This allows for archiving the todo, but keeping the
finished task.