Add section for standalone tasks to agenda view
Helper functions needed to be added in order to make it possible to
differentiate between normal tasks and projects (tasks with subtasks).
I documented those functions a little bit more then usual to make them
easier to understand.
The new section in the agenda view will show tasks that exist without
the context of a project.