Add project sections to agenda view
Shows stuck projects, projects, project next tasks, projects in planning
phase and task that are waiting in the agenda.
This is more or less borrowed from Bernt Hansen's setup with some
changes here and there.
I tried to document the functions needed for the view a little bit more
detailed than usual in order to make them understandable when revisiting
the code.