Projects with a scheduled call are not stuck
If a projects only has a scheduled call left, it won't show up in the
stuck project agenda.
Also don't show calls in the project next task section.

Those changes are necessary because I moved the calls that are part of a
project into the projects hierarchy. Previously there was a separate
heading for scheduled calls and calls belonging to a project were
collected there.