Rework of contacts section incl. capture template
The properties of the capture template for contacts are updated to the
once I used in previous setups. It now has all of the most important
properties. Some automation is now present, e.g. for entering the
crucial properties BIRTHDAY and EMAIL, which break the agenda view if
not present for a contact.

The second big part of this commit are the custom functions to get a
link to a contact entry, display or insert a contacts property or update
a contacts property interactively.
This will make it straight forward to get contact information into
contacts.org and also get information out of it.